pyproject.toml
[project]
name = "ai-labs-assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix>=13.3.0",
    "easyocr>=1.7.2",
    "geopandas>=1.1.2",
    "geopy>=2.4.1",
    "gradio>=6.6.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.9",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "openinference-instrumentation-langchain>=0.1.59",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.14.3",
    "rasterio>=1.5.0",
    "scikit-learn>=1.8.0",
    "tqdm>=4.67.3",
    "xgboost>=3.2.0",
]

[dependency-groups]
dev = [
    "nbstripout>=0.9.0",
]
